bot/requirements.txt
aiogram==3.0.0b8
python-dotenv==1.0.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6